# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Bank System",
    "description": "API Bank System adalah sebuah antarmuka pemrograman aplikasi yang dirancang untuk mengelola dan memproses berbagai layanan perbankan secara efisien. API ini memungkinkan pengguna untuk melakukan transaksi seperti pengecekan saldo, transfer dana antar rekening, pembuatan akun baru, serta pengelolaan data nasabah. Dibangun dengan arsitektur yang fleksibel, API ini mendukung integrasi yang mudah dengan berbagai sistem eksternal, serta dilengkapi dengan fitur keamanan tingkat tinggi untuk menjaga kerahasiaan dan integritas data perbankan.\n\n API ini menggunakan metode standar HTTP seperti GET, POST, PUT, dan DELETE, dengan format respons berbasis JSON untuk memudahkan proses integrasi dan pemrosesan data.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "naufal.ndak@...."
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/transactions/{transactionId}": {
      "get": {
        "tags": ["Transaction"],
        "summary": "Get transaction by id",
        "description": "Returns a single pet",
        "operationId": "transactionId",
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "description": "ID of transaction to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID transaction"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
...
      "transaction": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "example": 500000,
            "description": "Amount of money transferred"
          },
          "source_account_id": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "description": "ID of the source bank account"
          },
          "destination_account_id": {
            "type": "integer",
            "format": "int64",
            "example": 2,
            "description": "ID of the destination bank account"
          },
          "sourceAccount": {
            "type": "object",
            "properties": {
              "account_id": {
                "type": "integer",
                "example": 1
              },
              "bank_name": {
                "type": "string",
                "example": "Bank ABC"
              },
              "bank_account_number": {
                "type": "string",
                "example": "1234567890"
              }
            },
            "description": "Details of the source account"
          },
          "destinationAccount": {
            "type": "object",
            "properties": {
              "account_id": {
                "type": "integer",
                "example": 2
              },
              "bank_name": {
                "type": "string",
                "example": "Bank XYZ"
              },
              "bank_account_number": {
                "type": "string",
                "example": "0987654321"
              }
            },
            "description": "Details of the destination account"
          }
        },
        "xml": {
          "name": "transaction"
        }
      },
...
    },
...
    "securitySchemes": {
      "accounttransaction_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://accounttransaction3.swagger.io/oauth/authorize",
            "scopes": {
              "write:accounts": "modify accounts in your account",
              "read:accounts": "read your accounts"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  },
...
}